test(parse): Split into dedicated file
